* commit '601e1c3729a1ff7cb1ea4229dd7e9d8260ce2b5c':
  Fix IO_INPUT on 64-bit.

+++ rsThreadIO.cpp
@@ -213,7 +213,7 @@ bool ThreadIO::sendToClient(RsMessageToClientType cmdID, uint32_t usrID, const v
 
     //ALOGE("sendToClient %i %i %i", cmdID, usrID, (int)dataLen);
     ClientCmdHeader hdr;
-    hdr.bytes = dataLen;
+    hdr.bytes = (uint32_t)dataLen;
     hdr.cmdID = cmdID;
     hdr.userID = usrID;
